v1.3.3: Fix treemenu search not being enabled
Function exists prematurely before search gets enabled because the menu is closed, so container is too little. If we move the search enabling function after treemenu gets built then it would get executed. Also remove the warning of 0 container because it's not informative: it's expected when the treemenu is closed!

Fix for https://github.com/Gapminder/tools-page/issues/160