v1.46.1: Markercontrols — fix selectDialog appearing in a wrong place...
selectDialog would only show up correctly when clicking on the root-level items. in case of nested items it would be positioned off-screen, because it only takes the account item's offset relatively to the nearest parent.

i replace the offset-calculating formula to a more simple one, it uses different method with getBoundingClientRect() instead of offsetTop.

Adding position:relative on .vzb-dialog-content allows the floating select-dialog to scroll together with the container, thus appear glued to its item